backend/app/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
biopython==1.85
bleach==6.2.0
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
celery==5.3.4
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
emotion-model==0.0.6
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Mail==0.10.0
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
git-filter-repo==2.45.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.64.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.26.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
JPype1==1.5.0
jwt==1.3.1
kaggle==1.7.4.5
keras==2.11.0
kiwisolver==1.4.5
kombu==5.5.4
konlpy==0.6.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.0
namex==0.0.8
netifaces==0.10.6
networkx==3.3
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pdfminer==20191125
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
Pillow==10.0.1
pipeline==0.0.3
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.19.6
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydot==3.0.1
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.2
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.0.1
regex==2024.7.24
requests==2.32.2
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
socketio==0.2.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
strictyaml==1.7.3
sympy==1.13.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typer==0.12.5
typing_extensions==4.12.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.24.0
vine==5.1.0
virtualenv==20.26.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.3
whisper==1.1.10
wrapt==1.14.1
yarl==1.9.4
yelpapi==2.5.1
zipp==3.19.0
